Made requested updates
