+++ conftest.py
@@ -1,0 +1,8 @@
+# This is to easily import functions/ files from elsewhere in the project directly into testing files that are using
+# pytest
+
+
+# For example: to import my functions from root/backend/preprocessing.py into
+# root/tests/backend/test_preprocessing.py, I just have to add
+#     from backend.preprocessing import (function)
+# to the top of my test_preprocessing.py file
